refactor(smart): move all smart lib files to a `smart` subdirectory
