+++ demo-smart/index.js
@@ -1,6 +1,6 @@
 import '../src/components/cc-toaster/cc-toaster.js';
 import { addTranslations, setLanguage } from '../src/lib/i18n/i18n.js';
-import { updateRootContext } from '../src/lib/smart-manager.js';
+import { updateRootContext } from '../src/lib/smart/smart-manager.js';
 import { lang, translations } from '../src/translations/translations.en.js';
 
 addTranslations(lang, translations);
