+++ sandbox/forms/form-demo-with-smart-component.js
@@ -4,18 +4,18 @@ import '../../src/components/cc-button/cc-button.js';
 import '../../src/components/cc-input-text/cc-input-text.js';
 import '../../src/components/cc-notice/cc-notice.js';
 import '../../src/components/cc-smart-container/cc-smart-container.js';
-import { defineSmartComponent } from '../../src/lib/define-smart-component.js';
 import { dispatchCustomEvent } from '../../src/lib/events.js';
 import { FormErrorFocusController } from '../../src/lib/form/form-error-focus-controller.js';
 import { formSubmit } from '../../src/lib/form/form-submit-directive.js';
 import { notifySuccess } from '../../src/lib/notifications.js';
-import { updateRootContext } from '../../src/lib/smart-manager.js';
+import { defineSmartComponent } from '../../src/lib/smart/define-smart-component.js';
+import { updateRootContext } from '../../src/lib/smart/smart-manager.js';
 
 /**
  * @typedef {import('../../src/components/cc-input-text/cc-input-text.js').CcInputText} CcInputText
  * @typedef {import('../../src/lib/form/form.types.js').FormDataMap} FormDataMap
  * @typedef {import('../../src/lib/send-to-api.types.js').ApiConfig} ApiConfig
- * @typedef {import('../../src/lib/smart-component.types.js').OnContextUpdateArgs<FormDemoWithSmartComponent>} OnContextUpdateArgs
+ * @typedef {import('../../src/lib/smart/smart-component.types.d.ts').OnContextUpdateArgs<FormDemoWithSmartComponent>} OnContextUpdateArgs
  * @typedef {import('./form-demo-with-smart-component.types.js').FormDemoWithSmartComponentState} FormDemoWithSmartComponentState
  * @typedef {import('lit/directives/ref.js').Ref<HTMLFormElement>} HTMLFormElementRef
  */
