+++ src/components/cc-env-var-form/cc-env-var-form.smart-env-var-addon.js
@@ -1,7 +1,7 @@
 // @ts-expect-error FIXME: remove when clever-client exports types
 import { getAllEnvVars } from '@clevercloud/client/esm/api/v2/addon.js';
-import { defineSmartComponent } from '../../lib/define-smart-component.js';
 import { sendToApi } from '../../lib/send-to-api.js';
+import { defineSmartComponent } from '../../lib/smart/define-smart-component.js';
 import '../cc-smart-container/cc-smart-container.js';
 import './cc-env-var-form.js';
 
@@ -9,7 +9,7 @@ import './cc-env-var-form.js';
  * @typedef {import('./cc-env-var-form.js').CcEnvVarForm} CcEnvVarForm
  * @typedef {import('../common.types.js').EnvVar} EnvVar
  * @typedef {import('../../lib/send-to-api.js').ApiConfig} ApiConfig
- * @typedef {import('../../lib/smart-component.types.js').OnContextUpdateArgs<CcEnvVarForm>} OnContextUpdateArgs
+ * @typedef {import('../../lib/smart/smart-component.types.d.ts').OnContextUpdateArgs<CcEnvVarForm>} OnContextUpdateArgs
  */
 
 defineSmartComponent({
