+++ src/components/cc-jenkins-info/cc-jenkins-info.smart.js
@@ -1,7 +1,7 @@
 // @ts-expect-error FIXME: remove when clever-client exports types
 import { getAddon, getJenkinsUpdates } from '@clevercloud/client/esm/api/v4/addon-providers.js';
-import { defineSmartComponent } from '../../lib/define-smart-component.js';
 import { sendToApi } from '../../lib/send-to-api.js';
+import { defineSmartComponent } from '../../lib/smart/define-smart-component.js';
 import '../cc-smart-container/cc-smart-container.js';
 import './cc-jenkins-info.js';
 
@@ -9,7 +9,7 @@ import './cc-jenkins-info.js';
  * @typedef {import('./cc-jenkins-info.js').CcJenkinsInfo} CcJenkinsInfo
  * @typedef {import('./cc-jenkins-info.types.js').JenkinsInfoStateLoaded} JenkinsInfoStateLoaded
  * @typedef {import('../../lib/send-to-api.js').ApiConfig} ApiConfig
- * @typedef {import('../../lib/smart-component.types.js').OnContextUpdateArgs<CcJenkinsInfo>} OnContextUpdateArgs
+ * @typedef {import('../../lib/smart/smart-component.types.d.ts').OnContextUpdateArgs<CcJenkinsInfo>} OnContextUpdateArgs
  */
 
 defineSmartComponent({
