+++ src/components/cc-logs-application-view/cc-logs-application-view.smart.js
@@ -3,8 +3,8 @@ import { pickNonNull } from '@clevercloud/client/esm/pick-non-null.js';
 import { ApplicationLogStream } from '@clevercloud/client/esm/streams/application-logs.js';
 import { HttpError } from '@clevercloud/client/esm/streams/clever-cloud-sse.js';
 import { Buffer } from '../../lib/buffer.js';
-import { defineSmartComponent } from '../../lib/define-smart-component.js';
 import { sendToApi } from '../../lib/send-to-api.js';
+import { defineSmartComponent } from '../../lib/smart/define-smart-component.js';
 import { unique } from '../../lib/utils.js';
 import '../cc-smart-container/cc-smart-container.js';
 import './cc-logs-application-view.js';
@@ -20,7 +20,7 @@ import { isLive, lastXDays } from './date-range.js';
  * @typedef {import('./cc-logs-application-view.types.js').DateRange} DateRange
  * @typedef {import('./cc-logs-application-view.types.js').DateRangeSelection} DateRangeSelection
  * @typedef {import('./cc-logs-application-view.js').CcLogsApplicationView} CcLogsApplicationView
- * @typedef {import('../../lib/smart-component.types.js').OnContextUpdateArgs<CcLogsApplicationView>} OnContextUpdateArgs
+ * @typedef {import('../../lib/smart/smart-component.types.d.ts').OnContextUpdateArgs<CcLogsApplicationView>} OnContextUpdateArgs
  */
 
 /**
