+++ src/components/cc-pricing-product/cc-pricing-product.smart-addon.js
@@ -3,9 +3,9 @@ import { getAllAddonProviders } from '@clevercloud/client/esm/api/v2/product.js'
 // @ts-expect-error FIXME: remove when clever-client exports types
 import { ONE_DAY } from '@clevercloud/client/esm/with-cache.js';
 import { fetchPriceSystem } from '../../lib/api-helpers.js';
-import { defineSmartComponent } from '../../lib/define-smart-component.js';
 import { formatAddonProduct } from '../../lib/product.js';
 import { sendToApi } from '../../lib/send-to-api.js';
+import { defineSmartComponent } from '../../lib/smart/define-smart-component.js';
 import '../cc-smart-container/cc-smart-container.js';
 import './cc-pricing-product.js';
 
@@ -17,7 +17,7 @@ import './cc-pricing-product.js';
  * @typedef {import('../common.types.js').Zone} Zone
  * @typedef {import('../common.types.js').Instance} Instance
  * @typedef {import('../../lib/send-to-api.types.js').ApiConfig} ApiConfig
- * @typedef {import('../../lib/smart-component.types.js').OnContextUpdateArgs<CcPricingProduct>} OnContextUpdateArgs
+ * @typedef {import('../../lib/smart/smart-component.types.d.ts').OnContextUpdateArgs<CcPricingProduct>} OnContextUpdateArgs
  */
 
 defineSmartComponent({
