+++ src/components/cc-tile-metrics/cc-tile-metrics.smart.js
@@ -1,6 +1,6 @@
 import { getAppMetrics, getGrafanaOrganisation } from '../../lib/api-helpers.js';
-import { defineSmartComponent } from '../../lib/define-smart-component.js';
 import { sendToApi } from '../../lib/send-to-api.js';
+import { defineSmartComponent } from '../../lib/smart/define-smart-component.js';
 import '../cc-smart-container/cc-smart-container.js';
 import './cc-tile-metrics.js';
 
@@ -10,7 +10,7 @@ import './cc-tile-metrics.js';
  * @typedef {import('./cc-tile-metrics.types.js').MetricsData} MetricsData
  * @typedef {import('./cc-tile-metrics.types.js').Metric} Metric
  * @typedef {import('../../lib/send-to-api.types.js').ApiConfig} ApiConfig
- * @typedef {import('../../lib/smart-component.types.js').OnContextUpdateArgs<CcTileMetrics>} OnContextUpdateArgs
+ * @typedef {import('../../lib/smart/smart-component.types.d.ts').OnContextUpdateArgs<CcTileMetrics>} OnContextUpdateArgs
  */
 
 const NUMBER_OF_POINTS = 24;
