+++ src/components/cc-tile-status-codes/cc-tile-status-codes.smart.js
@@ -6,8 +6,8 @@ import { getWarp10AccessLogsToken } from '@clevercloud/client/esm/api/v2/warp-10
 import { THIRTY_SECONDS } from '@clevercloud/client/esm/request.fetch-with-timeout.js';
 // @ts-expect-error FIXME: remove when clever-client exports types
 import { ONE_DAY } from '@clevercloud/client/esm/with-cache.js';
-import { defineSmartComponent } from '../../lib/define-smart-component.js';
 import { sendToApi, sendToWarp } from '../../lib/send-to-api.js';
+import { defineSmartComponent } from '../../lib/smart/define-smart-component.js';
 import '../cc-smart-container/cc-smart-container.js';
 import './cc-tile-status-codes.js';
 
@@ -16,7 +16,7 @@ import './cc-tile-status-codes.js';
  * @typedef {import('./cc-tile-status-codes.types.js').StatusCodesData} StatusCodesData
  * @typedef {import('../../lib/send-to-api.types.js').ApiConfig} ApiConfig
  * @typedef {import('../../lib/send-to-api.types.js').Warp10ApiConfig} Warp10ApiConfig
- * @typedef {import('../../lib/smart-component.types.js').OnContextUpdateArgs<CcTileStatusCodes>} OnContextUpdateArgs
+ * @typedef {import('../../lib/smart/smart-component.types.d.ts').OnContextUpdateArgs<CcTileStatusCodes>} OnContextUpdateArgs
  */
 
 defineSmartComponent({
