+++ src/lib/smart/smart-component.types.d.ts
