+++ src/lib/smart/smart-symbols.js
